backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
redis>=5.0.0
playwright>=1.41.0
google-generativeai>=0.4.0
pydantic>=2.5.0
python-multipart>=0.0.6
